+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -655,5 +655,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-05T19:58:06.447Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-05T20:06:05.199Z")
     r"""The date and time the resource was last updated"""