+++ packages/orq-rc/RELEASES.md
@@ -2719,3 +2719,13 @@ Based on:
 - [python v3.12.0-rc.49] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.49] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.49 - packages/orq-rc
+
+## 2025-09-05 20:07:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.611.0 (2.694.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.50] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.50] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.50 - packages/orq-rc